requirements.txt
jupyter==1.0.0
pandas==1.0.3
python-dotenv==0.13.0
imageio==2.8.0
